+++ modules/flowable-engine-common/src/main/java/org/flowable/common/engine/impl/variable/VariableCollectionsContainerImpl.java
@@ -10,7 +10,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.flowable.common.engine.impl;
+package org.flowable.common.engine.impl.variable;
 
 import java.util.HashMap;
 import java.util.Map;
