+++ modules/flowable-engine/src/main/java/org/flowable/engine/impl/TaskCompletionBuilderImpl.java
@@ -19,7 +19,7 @@ import org.flowable.engine.impl.cmd.CompleteTaskCmd;
 import org.flowable.engine.impl.cmd.CompleteTaskWithFormCmd;
 import org.flowable.task.api.TaskCompletionBuilder;
 import org.flowable.common.engine.api.variable.VariableCollectionsContainer;
-import org.flowable.common.engine.impl.VariableCollectionsContainerImpl;
+import org.flowable.common.engine.impl.variable.VariableCollectionsContainerImpl;
 
 
 /**
