+++ modules/flowable-engine/src/test/java/org/flowable/standalone/escapeclause/HistoricTaskQueryEscapeClauseTest.java
@@ -73,54 +73,36 @@ public class HistoricTaskQueryEscapeClauseTest extends AbstractEscapeClauseTestC
         Map<String, Object> vars2 = new HashMap<>();
         vars2.put("var1", "Two_");
 
-
-
-        // VariableCollectionsContainer variableCollectionsContainerOne = new VariableCollectionsContainerImpl();
-        // variableCollectionsContainerOne.setLocalVariables(vars1);
-
         task1 = taskService.createTaskQuery().processInstanceId(processInstance1.getId()).singleResult();
         taskService.setAssignee(task1.getId(), "assignee%");
         taskService.setOwner(task1.getId(), "owner%");
-        // taskService.complete(task1.getId(), variableCollectionsContainerOne.getVariablesLocal(),true);
 
         TaskCompletionBuilder taskCompletionBuilder = taskService.createTaskCompletionBuilder();
         taskCompletionBuilder.variablesLocal(vars1);
         taskCompletionBuilder.taskId(task1.getId());
         taskCompletionBuilder.complete();
 
-        // VariableCollectionsContainer variableCollectionsContainerTwo = new VariableCollectionsContainerImpl();
-        // variableCollectionsContainerTwo.setLocalVariables(vars2);
-
         task2 = taskService.createTaskQuery().processInstanceId(processInstance1.getId()).singleResult();
         taskService.setAssignee(task2.getId(), "assignee_");
         taskService.setOwner(task2.getId(), "owner_");
-        //taskService.complete(task2.getId(), variableCollectionsContainerTwo.getVariablesLocal(),true);
 
         taskCompletionBuilder = taskService.createTaskCompletionBuilder();
         taskCompletionBuilder.variablesLocal(vars2);
         taskCompletionBuilder.taskId(task2.getId());
         taskCompletionBuilder.complete();
 
-        // VariableCollectionsContainer variableCollectionsContainerThree = new VariableCollectionsContainerImpl();
-        // variableCollectionsContainerThree.setLocalVariables(vars1);
-
         task3 = taskService.createTaskQuery().processInstanceId(processInstance2.getId()).singleResult();
         taskService.setAssignee(task3.getId(), "assignee%");
         taskService.setOwner(task3.getId(), "owner%");
-        // taskService.complete(task3.getId(), variableCollectionsContainerThree.getVariablesLocal(),true);
 
         taskCompletionBuilder = taskService.createTaskCompletionBuilder();
         taskCompletionBuilder.variablesLocal(vars1);
         taskCompletionBuilder.taskId(task3.getId());
         taskCompletionBuilder.complete();
 
-        // VariableCollectionsContainer variableCollectionsContainerFour = new VariableCollectionsContainerImpl();
-        // variableCollectionsContainerFour.setLocalVariables(vars2);
-
         task4 = taskService.createTaskQuery().processInstanceId(processInstance2.getId()).singleResult();
         taskService.setAssignee(task4.getId(), "assignee_");
         taskService.setOwner(task4.getId(), "owner_");
-        // taskService.complete(task4.getId(), variableCollectionsContainerFour.getVariablesLocal(),true);
 
         taskCompletionBuilder = taskService.createTaskCompletionBuilder();
         taskCompletionBuilder.variablesLocal(vars2);
